move all factories into one class
